Migrate remaining ui_panels code to src
