+++ __init__.py
@@ -39,7 +39,6 @@ from . import (  # noqa: WPS235
     tasks_queue,
     thread_tools,
     ui,
-    ui_panels,
     upload,
     utils,
 )
@@ -47,6 +46,7 @@ from .config import HANA3D_DESCRIPTION, HANA3D_NAME, HANA3D_UI
 from .src import async_loop, autothumb
 from .src.application.application import Application
 from .src.authentication.authentication import Authentication
+from .src.panels import panel_builder
 
 bl_info = {
     'name': 'Hana3D',
@@ -315,7 +315,7 @@ modules = (
     thread_tools,
     hana3d_types,
     ui,
-    ui_panels,
+    panel_builder,
     upload,
 )
