--- src/panels/builder.py
@@ -1,1 +1,0 @@
-"""Panel builder."""
